updated with example 6
Example 6 has a much improved way of handling vectors. This is done by
utilizing the vector object that is built in to numpy. Didn't know it
existed for the other examples.

I've also cleaned up the code a little so the example numbers are at the
top of each code block. I'll follow this convention going forward.